supervisor-frontend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "supervisor-frontend"
version = "0.1.0"
authors = [
  { name="Your Name", email="your.email@example.com" },
]
description = "Supervisor dashboard for trading bot"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "streamlit",
    "pandas",
    "numpy",
    "pyzmq",
    "plotly",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/trading-bot"
